feat(ci): Add tests job to build-terragraph-image-x86 GitHub Actions Workflow
The tests job runs the test executables shipped with the
terragraph-image-x86 target. The tests are located at `/usr/sbin/tests`
in the rootfs of terragraph-image-x86. The workflow fails if any of the
tests exit with a non-zero exit code.

`update_firewall` is removed from the rootfs prior to testing. This
executable is not expected to function in the terragraph-image-x86
chroot enviroment, and the presence of `update_firewall` will cause the
`tests` job to fail erroneously.

Some whitespace is also removed from build-terragraph-image-x86.yml.
`rm_work` is added to the `build` step of the workflow to preserve disk
space.